moved config and fixed imports
