+++ flow/config/json/emulator.go
@@ -19,7 +19,7 @@
 package json
 
 import (
-	"github.com/onflow/flow-cli/flow/project/cli/config"
+	"github.com/onflow/flow-cli/flow/config"
 )
 
 type jsonEmulators map[string]jsonEmulator
